Move `TryErr` into `Matches` lint pass
